requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
dotenv-python==0.0.1
fonttools==4.59.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kagglehub==0.3.12
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
ollama==0.5.3
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
